Level now actually loads level based on enum

+++ Game Code/Headers/checkPos.h
@@ -14,7 +14,7 @@ public:
 	//Overriding Methods
 	virtual void Update(sf::Time _frameTime);
 
-	void setText(sf::Vector2i _playerPos);
+	void setText(sf::Vector2f _playerPos);
 
 
 private:
